feat: Move typings/overrides to app folder
