+++ packages/core/app/components/sections/Alert/OverriddenDefaultAlert.ts
@@ -1,10 +1,10 @@
 'use client'
 
-import { override } from 'src/customizations/src/components/overrides/Alert'
 import { getOverriddenSection } from 'app/sdk/overrides/getOverriddenSection'
+import { override } from 'src/customizations/src/components/overrides/Alert'
 import Alert from '.'
 
-import type { SectionOverrideDefinitionV1 } from 'src/typings/overridesDefinition'
+import type { SectionOverrideDefinitionV1 } from 'app/typings/overridesDefinition'
 
 /**
  * This component exists to support overrides 1.0
