+++ packages/core/app/components/sections/Hero/OverriddenDefaultHero.ts
@@ -4,7 +4,7 @@ import { getOverriddenSection } from 'app/sdk/overrides/getOverriddenSection'
 import { override } from 'src/customizations/src/components/overrides/Hero'
 import Hero from '.'
 
-import type { SectionOverrideDefinitionV1 } from 'src/typings/overridesDefinition'
+import type { SectionOverrideDefinitionV1 } from 'app/typings/overridesDefinition'
 
 /**
  * This component exists to support overrides 1.0
