+++ packages/core/app/components/sections/Navbar/OverriddenDefaultNavbar.ts
@@ -1,8 +1,8 @@
 'use client'
 
-import { override } from 'src/customizations/src/components/overrides/Navbar'
 import { getOverriddenSection } from 'app/sdk/overrides/getOverriddenSection'
-import type { SectionOverrideDefinitionV1 } from 'src/typings/overridesDefinition'
+import type { SectionOverrideDefinitionV1 } from 'app/typings/overridesDefinition'
+import { override } from 'src/customizations/src/components/overrides/Navbar'
 import Navbar from './Navbar'
 
 /**
