+++ packages/core/app/components/sections/RegionBar/OverriddenDefaultRegionBar.ts
@@ -1,6 +1,6 @@
-import { override } from 'src/customizations/src/components/overrides/RegionBar'
 import { getOverriddenSection } from 'app/sdk/overrides/getOverriddenSection'
-import type { SectionOverrideDefinitionV1 } from 'src/typings/overridesDefinition'
+import type { SectionOverrideDefinitionV1 } from 'app/typings/overridesDefinition'
+import { override } from 'src/customizations/src/components/overrides/RegionBar'
 import RegionBar from '.'
 
 /**
