+++ packages/core/index.ts
@@ -1,26 +1,26 @@
-export * from './src/sdk/overrides/PageProvider'
+export * from './app/sdk/overrides/PageProvider'
 
 export type {
   PDPContext,
   PLPContext,
   SearchPageContext,
-} from './src/sdk/overrides/PageProvider'
+} from './app/sdk/overrides/PageProvider'
 
 export { useProductsQuery } from './src/sdk/product/useProductsQuery'
 
-export * from './src/typings/overrides'
-export { getOverriddenSection } from './src/sdk/overrides/getOverriddenSection'
+export { getOverriddenSection } from './app/sdk/overrides/getOverriddenSection'
+export * from './app/typings/overrides'
 
 // Overridable Sections
 export { default as AlertSection } from './app/components/sections/Alert'
+export { default as HeroSection } from './app/components/sections/Hero'
+export { default as NavbarSection } from './app/components/sections/Navbar'
+export { default as RegionBarSection } from './app/components/sections/RegionBar'
 export { default as BannerTextSection } from './src/components/sections/BannerText'
 export { default as BreadcrumbSection } from './src/components/sections/Breadcrumb'
 export { default as CrossSellingShelfSection } from './src/components/sections/CrossSellingShelf'
 export { default as EmptyState } from './src/components/sections/EmptyState'
-export { default as HeroSection } from './src/components/sections/Hero'
-export { default as NavbarSection } from './app/components/sections/Navbar'
 export { default as NewsletterSection } from './src/components/sections/Newsletter'
 export { default as ProductDetailsSection } from './src/components/sections/ProductDetails'
 export { default as ProductGallerySection } from './src/components/sections/ProductGallery'
 export { default as ProductShelfSection } from './src/components/sections/ProductShelf'
-export { default as RegionBarSection } from './app/components/sections/RegionBar'
