+++ packages/core/src/components/sections/Breadcrumb/OverriddenDefaultBreadcrumb.ts
@@ -1,8 +1,8 @@
-import { override } from 'src/customizations/src/components/overrides/Breadcrumb'
 import { getOverriddenSection } from 'app/sdk/overrides/getOverriddenSection'
+import { override } from 'src/customizations/src/components/overrides/Breadcrumb'
 import Breadcrumb from '.'
 
-import type { SectionOverrideDefinitionV1 } from 'src/typings/overridesDefinition'
+import type { SectionOverrideDefinitionV1 } from 'app/typings/overridesDefinition'
 
 /**
  * This component exists to support overrides 1.0
