+++ packages/core/src/components/sections/CrossSellingShelf/OverriddenDefaultCrossSellingShelf.ts
@@ -1,6 +1,6 @@
-import { override } from 'src/customizations/src/components/overrides/CrossSellingShelf'
 import { getOverriddenSection } from 'app/sdk/overrides/getOverriddenSection'
-import type { SectionOverrideDefinitionV1 } from 'src/typings/overridesDefinition'
+import type { SectionOverrideDefinitionV1 } from 'app/typings/overridesDefinition'
+import { override } from 'src/customizations/src/components/overrides/CrossSellingShelf'
 import CrossSellingShelf from '.'
 
 /**
