+++ packages/core/src/components/sections/ProductDetails/OverriddenDefaultProductDetails.ts
@@ -1,8 +1,8 @@
-import { override } from 'src/customizations/src/components/overrides/ProductDetails'
 import { getOverriddenSection } from 'app/sdk/overrides/getOverriddenSection'
+import { override } from 'src/customizations/src/components/overrides/ProductDetails'
 import ProductDetails from './ProductDetails'
 
-import type { SectionOverrideDefinitionV1 } from 'src/typings/overridesDefinition'
+import type { SectionOverrideDefinitionV1 } from 'app/typings/overridesDefinition'
 
 /**
  * This component exists to support overrides 1.0
