+++ packages/core/src/customizations/src/components/overrides/EmptyState.tsx
@@ -1,5 +1,5 @@
 // This is an example of how it can be used on the starter.
-import { SectionOverride } from 'src/typings/overrides'
+import { SectionOverride } from 'app/typings/overrides'
 
 const SECTION = 'EmptyState' as const
